requirements.txt
#wheel
#bs4
#cltk
#keras
#sklearn_crfsuite
#progress

# lsnn
#scikit-learn
#tensorflow
#tensorflow-addons
#numpy
#pandas
#unidecode
#seaborn 

# Hypotactic
selenium
bs4

# Pedecerto
ipython
lxml
mqdqparser
